# openapi.yml
openapi: 3.0.3
info:
  title: QTYM Backend API
  version: 1.0.0
  description: Quantity measurement backend for user auth and unit conversion operations.
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Health
  - name: Users
  - name: Conversions
paths:
  /:
    get:
      tags: [Health]
      summary: Root health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/user/health:
    get:
      tags: [Health]
      summary: User service health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/user/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/user:
    get:
      tags: [Users]
      summary: List all users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/ValidationError'
        "409":
          $ref: '#/components/responses/ConflictError'
        "500":
          $ref: '#/components/responses/ServerError'
  /api/v1/user/login:
    post:
      tags: [Users]
      summary: Login a user
      description: Implemented as GET in the current codebase and reads body fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          $ref: '#/components/responses/ServerError'
  /api/v1/conversions/health:
    get:
      tags: [Health]
      summary: Conversion service health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/conversions/convert:
    post:
      tags: [Conversions]
      summary: Convert one unit quantity to another unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertRequest'
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        "400":
          $ref: '#/components/responses/BadRequestError'
  /api/v1/conversions/add:
    post:
      tags: [Conversions]
      summary: Add two quantities and normalize them to one unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRequest'
      responses:
        "200":
          description: Addition result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        "400":
          $ref: '#/components/responses/BadRequestError'
  /api/v1/conversions/compare:
    post:
      tags: [Conversions]
      summary: Compare two quantities of different units after normalization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        "200":
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        "400":
          $ref: '#/components/responses/BadRequestError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/ValidationItem'
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationItem:
      type: object
      properties:
        type:
          type: string
        value:
          nullable: true
        msg:
          type: string
        path:
          type: string
        location:
          type: string
      required: [msg, path, location]
    Unit:
      type: object
      properties:
        unitType:
          type: string
          example: LENGTH
        label:
          type: string
          example: Meter
        symbol:
          type: string
          example: m
      required: [unitType, label, symbol]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
      required: [name, email, phoneNumber, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        token:
          type: string
          description: JWT used in the Swagger Authorize dialog
        user:
          $ref: '#/components/schemas/User'
      required: [message, token, user]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        passwordHash:
          type: string
      required: [id, name, phoneNumber, email, passwordHash]
    ConvertRequest:
      type: object
      properties:
        fromQty:
          type: number
        fromUnit:
          $ref: '#/components/schemas/Unit'
        toUnit:
          $ref: '#/components/schemas/Unit'
      required: [fromQty, fromUnit, toUnit]
    AddRequest:
      type: object
      properties:
        fromQty:
          type: number
        fromUnit:
          $ref: '#/components/schemas/Unit'
        toQty:
          type: number
        toUnit:
          $ref: '#/components/schemas/Unit'
        normalizeUnit:
          $ref: '#/components/schemas/Unit'
      required: [fromQty, fromUnit, toQty, toUnit]
    CompareRequest:
      type: object
      properties:
        fromQty:
          type: number
        fromUnit:
          $ref: '#/components/schemas/Unit'
        toQty:
          type: number
        toUnit:
          $ref: '#/components/schemas/Unit'
        normalizeUnit:
          $ref: '#/components/schemas/Unit'
      required: [fromQty, fromUnit, toQty, toUnit]
    HistoryResponse:
      type: object
      properties:
        fromValue:
          type: number
        fromUnit:
          $ref: '#/components/schemas/Unit'
        toValue:
          type: number
        toUnit:
          $ref: '#/components/schemas/Unit'
        resultVal:
          type: number
        operationResult:
          type: string
        operationType:
          type: string
          enum:
            - CONVERSION
            - COMPARISON
            - ARITHMETIC_ADD
            - ARITHMETIC_SUBTRACT
            - ARITHMETIC_DIVIDE
        normalizedUnit:
          $ref: '#/components/schemas/Unit'
      required:
        - fromValue
        - fromUnit
        - toValue
        - toUnit
        - resultVal
        - operationResult
        - operationType
        - normalizedUnit